Break into a file per flight
