--- internal/flight/flight13/flighthandlers_client.go
@@ -1,889 +1,0 @@
-// SPDX-FileCopyrightText: 2026 The Pion community <https://pion.ly>
-// SPDX-License-Identifier: MIT
-
-package flight13
-
-import (
-	"bytes"
-	"context"
-	"crypto"
-	"crypto/tls"
-	"errors"
-	"maps"
-	"slices"
-
-	"github.com/pion/dtls/v3/internal/ciphersuite"
-	dtlsconfig "github.com/pion/dtls/v3/internal/config"
-	dtlserrors "github.com/pion/dtls/v3/internal/errors"
-	dtlsflight "github.com/pion/dtls/v3/internal/flight"
-	"github.com/pion/dtls/v3/pkg/crypto/elliptic"
-	"github.com/pion/dtls/v3/pkg/crypto/prf"
-	"github.com/pion/dtls/v3/pkg/crypto/signaturehash"
-	"github.com/pion/dtls/v3/pkg/protocol"
-	"github.com/pion/dtls/v3/pkg/protocol/alert"
-	"github.com/pion/dtls/v3/pkg/protocol/extension"
-	"github.com/pion/dtls/v3/pkg/protocol/handshake"
-	"github.com/pion/dtls/v3/pkg/protocol/recordlayer"
-)
-
-// we'll add the flight handlers for the DTLS 1.3 client here.
-//
-// +----------+
-// | Flight 1 |
-// | Flight 3 |
-// | Flight 5 |
-// +----------+
-//
-// +-----------+
-// | Flight 3a |
-// | Flight 5a |
-// +-----------+
-//
-// +-----------+
-// | Flight 3b |
-// | Flight 5b |
-// +-----------+
-//
-// +-----------+
-// | Flight 5c |
-// +-----------+
-
-type serverHelloPull struct {
-	nextHandshakeSequence int
-	serverHello           *handshake.MessageServerHello
-	items                 []*dtlsflight.HandshakeCacheItem
-	ready                 bool
-	failure               *flightParseFailure
-}
-
-func IsHelloRetryRequest(sh *handshake.MessageServerHello) bool {
-	randomBytes := sh.Random.MarshalFixed()
-
-	return bytes.Equal(randomBytes[:], handshake.HelloRetryRequestRandom())
-}
-
-func ServerHelloSelectedVersions(extensions []extension.Extension) ([]protocol.Version, bool, error) {
-	seenSupportedVersions := false
-	var versions []protocol.Version
-	for _, val := range extensions {
-		supportedVersions, ok := val.(*extension.SupportedVersions)
-		if !ok {
-			continue
-		}
-		if seenSupportedVersions || !supportedVersions.IsSelectedVersion() || len(supportedVersions.Versions) != 1 {
-			return nil, true, dtlserrors.ErrInvalidServerHello
-		}
-		seenSupportedVersions = true
-		versions = supportedVersions.Versions
-	}
-
-	return versions, seenSupportedVersions, nil
-}
-
-func validateHelloRetryRequestSelectedVersion(extensions []extension.Extension) error {
-	versions, seenSupportedVersions, err := ServerHelloSelectedVersions(extensions)
-	if err != nil || !seenSupportedVersions {
-		return dtlserrors.ErrInvalidHelloRetryRequest
-	}
-	if !versions[0].Equal(protocol.Version1_3) {
-		return dtlserrors.ErrUnsupportedProtocolVersion
-	}
-
-	return nil
-}
-
-func selectServerHelloCipherSuite(
-	serverHello *handshake.MessageServerHello,
-	cfg *dtlsconfig.HandshakeConfig,
-) (dtlsconfig.CipherSuite, *alert.Alert, error) {
-	if serverHello.CipherSuiteID == nil {
-		return nil, &alert.Alert{Level: alert.Fatal, Description: alert.IllegalParameter},
-			dtlserrors.ErrInvalidServerHello
-	}
-	remoteCipherSuite := ciphersuite.ForID(ciphersuite.ID(*serverHello.CipherSuiteID), cfg.CustomCipherSuites)
-	if remoteCipherSuite == nil {
-		return nil, &alert.Alert{Level: alert.Fatal, Description: alert.InsufficientSecurity},
-			dtlserrors.ErrCipherSuiteNoIntersection
-	}
-	if !ciphersuite.IDSupportsVersion(remoteCipherSuite.ID(), protocol.Version1_3) {
-		return nil, &alert.Alert{Level: alert.Fatal, Description: alert.InsufficientSecurity},
-			dtlserrors.ErrInvalidCipherSuite
-	}
-	selectedCipherSuite, found := dtlsflight.FindMatchingCipherSuite(
-		[]dtlsconfig.CipherSuite{remoteCipherSuite}, cfg.LocalCipherSuites,
-	)
-	if !found {
-		return nil, &alert.Alert{Level: alert.Fatal, Description: alert.InsufficientSecurity},
-			dtlserrors.ErrInvalidCipherSuite
-	}
-
-	return selectedCipherSuite, nil, nil
-}
-
-// nolint:cyclop
-func flight1Parse(
-	ctx context.Context,
-	conn dtlsflight.Conn,
-	flightCtx *handshakeContext,
-) (Flight, *alert.Alert, error) {
-	state := flightCtx.state
-	cache := flightCtx.cache
-	cfg := flightCtx.cfg
-
-	seq, msgs, items, ok := cache.FullPullMapItems(state.HandshakeRecvSequence, state.CipherSuite,
-		dtlsflight.HandshakeCachePullRule{Typ: handshake.TypeServerHello, Epoch: cfg.InitialEpoch, IsClient: false, Optional: true}, //nolint:lll
-	)
-	if !ok {
-		// No valid message received. Keep reading
-		return 0, nil, nil
-	}
-
-	sh, ok := msgs[handshake.TypeServerHello].(*handshake.MessageServerHello)
-	if !ok {
-		return 0, &alert.Alert{Level: alert.Fatal, Description: alert.InternalError}, nil
-	}
-
-	if !IsHelloRetryRequest(sh) {
-		// Flight1 and flight2 were skipped.
-		// Parse as flight3.
-		return flight3Parse(ctx, conn, flightCtx)
-	}
-	// Handle HelloRetryRequest
-
-	if !sh.Version.Equal(protocol.Version1_0) && !sh.Version.Equal(protocol.Version1_2) {
-		return 0, &alert.Alert{Level: alert.Fatal, Description: alert.ProtocolVersion},
-			dtlserrors.ErrUnsupportedProtocolVersion
-	}
-	if err := validateHelloRetryRequestSelectedVersion(sh.Extensions); err != nil {
-		description := alert.IllegalParameter
-		if errors.Is(err, dtlserrors.ErrUnsupportedProtocolVersion) {
-			description = alert.ProtocolVersion
-		}
-
-		return 0, &alert.Alert{Level: alert.Fatal, Description: description}, err
-	}
-	selectedCipherSuite, dtlsAlert, err := selectServerHelloCipherSuite(sh, cfg)
-	if err != nil {
-		return 0, dtlsAlert, err
-	}
-	state.CipherSuite = selectedCipherSuite
-
-	// nolint:godox
-	// TODO: negotiate minimial set of extensions necessary for the client
-	// to generate a correct CH pair. As with the ServerHello, a
-	// HelloRetryRequest MUST NOT contain any extensions that were not first
-	// offered by the client in its ClientHello, with the exception of
-	// optionally the "cookie" extension
-	for _, val := range sh.Extensions {
-		switch ext := val.(type) {
-		case *extension.SupportedVersions:
-			// nolint:godox
-			// TODO: negotiate version
-			state.RemoteVersions = ext.Versions
-		case *extension.CookieExt:
-			state.Cookie = ext.Cookie
-		case *extension.KeyShare:
-			if ext.SelectedGroup != nil {
-				state.RemoteKeyEntries = &[]extension.KeyShareEntry{
-					{Group: *ext.SelectedGroup},
-				}
-			}
-		}
-	}
-
-	if flightCtx.inboundHandshakeHandler != nil {
-		if err := flightCtx.inboundHandshakeHandler(state.CipherSuite, items); err != nil {
-			return 0, &alert.Alert{Level: alert.Fatal, Description: alert.InternalError}, err
-		}
-	}
-	state.HandshakeRecvSequence = seq
-
-	return Flight3, nil, nil
-}
-
-func flight3Parse(
-	ctx context.Context,
-	conn dtlsflight.Conn,
-	flightCtx *handshakeContext,
-) (Flight, *alert.Alert, error) {
-	pull := flight3PullServerHello(flightCtx)
-	if !pull.ready {
-		return 0, nil, nil
-	}
-	if pull.failure != nil {
-		return 0, pull.failure.alert, pull.failure.err
-	}
-
-	failure := processFlight3ServerHello(flightCtx, pull.serverHello)
-	if failure != nil {
-		return 0, failure.alert, failure.err
-	}
-	failure = initializeFlight3HandshakeProtection(
-		ctx,
-		conn,
-		flightCtx,
-		pull.nextHandshakeSequence,
-		pull.items,
-	)
-	if failure != nil {
-		return 0, failure.alert, failure.err
-	}
-
-	protectedFlight := pullProtectedHandshakeFlight(
-		flightCtx.cache,
-		[]dtlsflight.HandshakeCachePullRule{
-			{Typ: handshake.TypeEncryptedExtensions, Epoch: EpochHandshake, IsClient: false, Optional: false},
-			{Typ: handshake.TypeCertificateRequest, Epoch: EpochHandshake, IsClient: false, Optional: true},
-			{Typ: handshake.TypeCertificate, Epoch: EpochHandshake, IsClient: false, Optional: true},
-			{Typ: handshake.TypeCertificateVerify, Epoch: EpochHandshake, IsClient: false, Optional: true},
-			{Typ: handshake.TypeFinished, Epoch: EpochHandshake, IsClient: false, Optional: false},
-		},
-		pull.nextHandshakeSequence,
-	)
-	if !protectedFlight.ready {
-		return 0, nil, nil
-	}
-	if protectedFlight.failure != nil {
-		return 0, protectedFlight.failure.alert, protectedFlight.failure.err
-	}
-	failure = handleFlight3ProtectedHandshake(flightCtx, protectedFlight.items)
-	if failure != nil {
-		return 0, failure.alert, failure.err
-	}
-	flightCtx.state.HandshakeRecvSequence = protectedFlight.nextHandshakeSequence
-
-	return Flight5, nil, nil
-}
-
-func flight3PullServerHello(
-	flightCtx *handshakeContext,
-) serverHelloPull {
-	seq, msgs, items, ok := flightCtx.cache.FullPullMapItems(
-		flightCtx.state.HandshakeRecvSequence, flightCtx.state.CipherSuite,
-		dtlsflight.HandshakeCachePullRule{Typ: handshake.TypeServerHello, Epoch: flightCtx.cfg.InitialEpoch, IsClient: false, Optional: false}, //nolint:lll
-	)
-	if !ok {
-		return serverHelloPull{}
-	}
-
-	serverHello, ok := msgs[handshake.TypeServerHello].(*handshake.MessageServerHello)
-	if !ok {
-		return serverHelloPull{
-			ready:   true,
-			failure: newFlightParseFailure(alert.InternalError, nil),
-		}
-	}
-
-	return serverHelloPull{
-		nextHandshakeSequence: seq,
-		serverHello:           serverHello,
-		items:                 items,
-		ready:                 true,
-	}
-}
-
-func processFlight3ServerHello(
-	flightCtx *handshakeContext,
-	serverHello *handshake.MessageServerHello,
-) *flightParseFailure {
-	versions, failure := validateFlight3ServerHello(serverHello)
-	if failure != nil {
-		return failure
-	}
-	flightCtx.state.RemoteVersions = versions
-	flightCtx.state.LocalVersion = protocol.Version1_3
-
-	selectedCipherSuite, dtlsAlert, err := selectServerHelloCipherSuite(serverHello, flightCtx.cfg)
-	if err != nil {
-		return &flightParseFailure{alert: dtlsAlert, err: err}
-	}
-	flightCtx.state.CipherSuite = selectedCipherSuite
-	flightCtx.state.RemoteRandom = serverHello.Random
-	flightCtx.cfg.Log.Tracef("[handshake13] use cipher suite: %s", selectedCipherSuite.String())
-
-	serverShare := serverHelloKeyShare(serverHello.Extensions)
-	if serverShare == nil {
-		return newFlightParseFailure(alert.IllegalParameter, dtlserrors.ErrServerKeyShareMissing)
-	}
-
-	return applyFlight3ServerKeyShare(flightCtx, serverShare)
-}
-
-func validateFlight3ServerHello(serverHello *handshake.MessageServerHello) ([]protocol.Version, *flightParseFailure) {
-	if IsHelloRetryRequest(serverHello) {
-		return nil, newFlightParseFailure(
-			alert.UnexpectedMessage,
-			dtlserrors.ErrUnexpectedSecondHelloRetryRequest,
-		)
-	}
-
-	if !serverHello.Version.Equal(protocol.Version1_2) {
-		return nil, newFlightParseFailure(alert.ProtocolVersion, dtlserrors.ErrUnsupportedProtocolVersion)
-	}
-
-	versions, seenSupportedVersions, err := ServerHelloSelectedVersions(serverHello.Extensions)
-	if err != nil {
-		return nil, newFlightParseFailure(alert.IllegalParameter, dtlserrors.ErrInvalidServerHello)
-	}
-	if !seenSupportedVersions || !versions[0].Equal(protocol.Version1_3) {
-		return nil, newFlightParseFailure(alert.ProtocolVersion, dtlserrors.ErrUnsupportedProtocolVersion)
-	}
-
-	return versions, nil
-}
-
-func serverHelloKeyShare(extensions []extension.Extension) *extension.KeyShareEntry {
-	for _, ext := range extensions {
-		keyShare, ok := ext.(*extension.KeyShare)
-		if !ok || keyShare.ServerShare == nil {
-			continue
-		}
-
-		return keyShare.ServerShare
-	}
-
-	return nil
-}
-
-func applyFlight3ServerKeyShare(
-	flightCtx *handshakeContext,
-	serverShare *extension.KeyShareEntry,
-) *flightParseFailure {
-	localKeypair, ok := flightCtx.state.LocalKeypairs[serverShare.Group]
-	if !ok || localKeypair == nil {
-		return newFlightParseFailure(alert.IllegalParameter, dtlserrors.ErrServerKeyShareUnknownGroup)
-	}
-
-	keyAgreementSecret, err := prf.PreMasterSecret(serverShare.KeyExchange, localKeypair.PrivateKey, serverShare.Group)
-	if err != nil {
-		return newFlightParseFailure(alert.InternalError, err)
-	}
-	flightCtx.state.KeyAgreementSecret = keyAgreementSecret
-	flightCtx.state.SelectedGroup = serverShare.Group
-	flightCtx.state.RemoteKeyEntries = &[]extension.KeyShareEntry{*serverShare}
-
-	return nil
-}
-
-func initializeFlight3HandshakeProtection(
-	ctx context.Context,
-	conn dtlsflight.Conn,
-	flightCtx *handshakeContext,
-	serverHelloSeq int,
-	items []*dtlsflight.HandshakeCacheItem,
-) *flightParseFailure {
-	if failure := handleFlight3InboundHandshake(flightCtx, items); failure != nil {
-		return failure
-	}
-	flightCtx.state.HandshakeRecvSequence = serverHelloSeq
-	if flightCtx.handshakeTrafficSecretDeriver != nil {
-		if err := flightCtx.handshakeTrafficSecretDeriver(flightCtx.state); err != nil {
-			return newFlightParseFailure(alert.InternalError, err)
-		}
-	}
-	if flightCtx.handshakeRecordProtectionInitializer == nil {
-		return nil
-	}
-	if err := flightCtx.handshakeRecordProtectionInitializer(flightCtx.state); err != nil {
-		return newFlightParseFailure(alert.InternalError, err)
-	}
-	flightCtx.state.RemoteEpoch.Store(EpochHandshake)
-	if conn == nil {
-		return nil
-	}
-	if err := conn.HandleQueuedPackets(ctx); err != nil {
-		return newFlightParseFailure(alert.InternalError, err)
-	}
-
-	return nil
-}
-
-func handleFlight3ProtectedHandshake(
-	flightCtx *handshakeContext,
-	items []*dtlsflight.HandshakeCacheItem,
-) *flightParseFailure {
-	if flightCtx.protectedHandshakeHandler == nil {
-		return newFlightParseFailure(alert.InternalError, dtlserrors.ErrHandshakeTranscriptHashNotSelected)
-	}
-	if err := flightCtx.protectedHandshakeHandler(flightCtx.state.CipherSuite, items); err != nil {
-		return protectedFlightParseFailure(err)
-	}
-
-	return nil
-}
-
-func protectedFlightParseFailure(err error) *flightParseFailure {
-	switch {
-	case errors.Is(err, dtlserrors.ErrVerifyDataMismatch):
-		return newFlightParseFailure(alert.HandshakeFailure, err)
-	case errors.Is(err, dtlserrors.ErrCertificateVerifyNoCertificate):
-		return newFlightParseFailure(alert.NoCertificate, err)
-	case errors.Is(err, dtlserrors.ErrClientCertificateRequired):
-		return newFlightParseFailure(alert.CertificateRequired, err)
-	case errors.Is(err, dtlserrors.ErrKeySignatureMismatch),
-		errors.Is(err, dtlserrors.ErrInvalidCertificate),
-		errors.Is(err, dtlserrors.ErrCertificateVerificationFailed),
-		errors.Is(err, dtlserrors.ErrClientCertificateNotVerified),
-		errors.Is(err, dtlserrors.ErrInvalidCertificateOID),
-		errors.Is(err, dtlserrors.ErrInvalidCertificateSignatureAlgorithm),
-		errors.Is(err, dtlserrors.ErrNotAcceptableCertificateChain):
-		return newFlightParseFailure(alert.BadCertificate, err)
-	case errors.Is(err, dtlserrors.ErrNoAvailableSignatureSchemes):
-		return newFlightParseFailure(alert.InsufficientSecurity, err)
-	default:
-		return newFlightParseFailure(alert.InternalError, err)
-	}
-}
-
-func handleFlight3InboundHandshake(
-	flightCtx *handshakeContext,
-	items []*dtlsflight.HandshakeCacheItem,
-) *flightParseFailure {
-	if flightCtx.inboundHandshakeHandler == nil {
-		return nil
-	}
-	if err := flightCtx.inboundHandshakeHandler(flightCtx.state.CipherSuite, items); err != nil {
-		return newFlightParseFailure(alert.InternalError, err)
-	}
-
-	return nil
-}
-
-//nolint:cyclop
-func flight1Generate(
-	_ dtlsflight.Conn,
-	flightCtx *handshakeContext,
-) ([]*dtlsflight.Packet, *alert.Alert, error) {
-	state := flightCtx.state
-	cfg := flightCtx.cfg
-
-	state.LocalEpoch.Store(EpochInitial)
-	state.RemoteEpoch.Store(EpochInitial)
-	if len(cfg.EllipticCurves) < 1 {
-		return nil, nil, dtlserrors.ErrEmptyEllipticCurves
-	}
-	if len(cfg.LocalSignatureSchemes) < 1 {
-		return nil, nil, dtlserrors.ErrNoAvailableSignatureSchemes
-	}
-	state.SelectedGroup = cfg.EllipticCurves[0]
-	state.Cookie = nil
-
-	if err := state.LocalRandom.Populate(); err != nil {
-		return nil, nil, err
-	}
-
-	if cfg.HelloRandomBytesGenerator != nil {
-		state.LocalRandom.RandomBytes = cfg.HelloRandomBytesGenerator()
-	}
-
-	extensions := []extension.Extension{
-		&extension.SupportedSignatureAlgorithms{
-			SignatureHashAlgorithms: cfg.LocalSignatureSchemes,
-		},
-	}
-
-	if cfg.ExtendedMasterSecret == dtlsconfig.RequestExtendedMasterSecret ||
-		cfg.ExtendedMasterSecret == dtlsconfig.RequireExtendedMasterSecret {
-		extensions = append(extensions, &extension.UseExtendedMasterSecret{
-			Supported: true,
-		})
-	}
-
-	extensions = append(extensions, &extension.RenegotiationInfo{
-		RenegotiatedConnection: 0,
-	})
-
-	var setEllipticCurveCryptographyClientHelloExtensions bool
-	for _, c := range cfg.LocalCipherSuites {
-		if c.ECC() {
-			setEllipticCurveCryptographyClientHelloExtensions = true
-
-			break
-		}
-	}
-
-	if setEllipticCurveCryptographyClientHelloExtensions {
-		extensions = append(extensions, []extension.Extension{
-			&extension.SupportedEllipticCurves{
-				EllipticCurves: cfg.EllipticCurves,
-			},
-			&extension.SupportedPointFormats{
-				PointFormats: []elliptic.CurvePointFormat{elliptic.CurvePointFormatUncompressed},
-			},
-		}...)
-	}
-
-	if len(cfg.SupportedProtocols) > 0 {
-		extensions = append(extensions, &extension.ALPN{ProtocolNameList: cfg.SupportedProtocols})
-	}
-
-	entries := make([]extension.KeyShareEntry, 0, len(cfg.EllipticCurves))
-	keypairs := make(map[elliptic.Curve]*elliptic.Keypair, len(cfg.EllipticCurves))
-	for _, group := range cfg.EllipticCurves {
-		keypair, err := elliptic.GenerateKeypair(group)
-		if err != nil {
-			return nil, nil, err
-		}
-		entries = append(entries, extension.KeyShareEntry{
-			Group: keypair.Curve, KeyExchange: keypair.PublicKey,
-		})
-		keypairs[keypair.Curve] = keypair
-	}
-	state.LocalKeyEntries = entries
-	state.LocalKeypairs = keypairs
-	extensions = append(extensions, &extension.KeyShare{
-		ClientShares: entries,
-	})
-
-	extensions = append(extensions, &extension.SupportedVersions{
-		Versions: dtlsconfig.SupportedVersionsRange(cfg.MinVersion, cfg.MaxVersion),
-	})
-
-	if len(cfg.LocalCertSignatureSchemes) > 0 {
-		extensions = append(extensions, &extension.SignatureAlgorithmsCert{
-			SignatureHashAlgorithms: cfg.LocalCertSignatureSchemes,
-		})
-	}
-
-	if len(cfg.ServerName) > 0 {
-		extensions = append(extensions, &extension.ServerName{ServerName: cfg.ServerName})
-	}
-
-	if len(cfg.LocalSRTPProtectionProfiles) > 0 {
-		extensions = append(extensions, &extension.UseSRTP{
-			ProtectionProfiles:  cfg.LocalSRTPProtectionProfiles,
-			MasterKeyIdentifier: cfg.LocalSRTPMasterKeyIdentifier,
-		})
-	}
-
-	// connection ID
-
-	// Pre_shared_key must be last extension
-
-	clientHello := &handshake.MessageClientHello{
-		Version:   protocol.Version1_2,
-		SessionID: state.SessionID,
-		Cookie:    nil,
-		Random:    state.LocalRandom,
-		// Add DTLS 1.3 ciphersuites
-		CipherSuiteIDs:     dtlsflight.CipherSuiteIDs(cfg.LocalCipherSuites),
-		CompressionMethods: dtlsflight.DefaultCompressionMethods(),
-		Extensions:         extensions,
-	}
-
-	var content handshake.Handshake
-
-	if cfg.ClientHelloMessageHook != nil {
-		content = handshake.Handshake{Message: cfg.ClientHelloMessageHook(*clientHello)}
-	} else {
-		content = handshake.Handshake{Message: clientHello}
-	}
-
-	return []*dtlsflight.Packet{
-		{
-			Record: &recordlayer.RecordLayer{
-				Header: recordlayer.Header{
-					Version: protocol.Version1_2,
-				},
-				Content: &content,
-			},
-		},
-	}, nil, nil
-}
-
-// nolint:cyclop
-func flight3Generate(
-	_ dtlsflight.Conn,
-	flightCtx *handshakeContext,
-) ([]*dtlsflight.Packet, *alert.Alert, error) {
-	if len(flightCtx.cfg.LocalSignatureSchemes) < 1 {
-		return nil, nil, dtlserrors.ErrNoAvailableSignatureSchemes
-	}
-
-	extensions := []extension.Extension{
-		&extension.SupportedSignatureAlgorithms{
-			SignatureHashAlgorithms: flightCtx.cfg.LocalSignatureSchemes,
-		},
-	}
-
-	if flightCtx.cfg.ExtendedMasterSecret == dtlsconfig.RequestExtendedMasterSecret ||
-		flightCtx.cfg.ExtendedMasterSecret == dtlsconfig.RequireExtendedMasterSecret {
-		extensions = append(extensions, &extension.UseExtendedMasterSecret{
-			Supported: true,
-		})
-	}
-
-	extensions = append(extensions, &extension.RenegotiationInfo{
-		RenegotiatedConnection: 0,
-	})
-
-	if flightCtx.state.SelectedGroup != 0 {
-		extensions = append(extensions, []extension.Extension{
-			&extension.SupportedEllipticCurves{
-				EllipticCurves: flightCtx.cfg.EllipticCurves,
-			},
-			&extension.SupportedPointFormats{
-				PointFormats: []elliptic.CurvePointFormat{elliptic.CurvePointFormatUncompressed},
-			},
-		}...)
-	}
-
-	if len(flightCtx.cfg.SupportedProtocols) > 0 {
-		extensions = append(extensions, &extension.ALPN{ProtocolNameList: flightCtx.cfg.SupportedProtocols})
-	}
-
-	var localGroups []elliptic.Curve
-	var newEntries []extension.KeyShareEntry
-	newKeypairs := map[elliptic.Curve]*elliptic.Keypair{}
-	if flightCtx.state.RemoteKeyEntries != nil {
-		for _, entry := range flightCtx.state.LocalKeyEntries {
-			localGroups = append(localGroups, entry.Group)
-		}
-
-		for _, entry := range *flightCtx.state.RemoteKeyEntries {
-			if !slices.Contains(localGroups, entry.Group) && slices.Contains(flightCtx.cfg.EllipticCurves, entry.Group) {
-				keypair, err := elliptic.GenerateKeypair(entry.Group)
-				if err != nil {
-					return nil, nil, err
-				}
-				newEntries = append(newEntries, extension.KeyShareEntry{
-					Group: keypair.Curve, KeyExchange: keypair.PublicKey,
-				})
-				newKeypairs[keypair.Curve] = keypair
-			}
-		}
-	}
-	if len(newEntries) > 0 {
-		flightCtx.state.LocalKeyEntries = append(newEntries, flightCtx.state.LocalKeyEntries...)
-		if flightCtx.state.LocalKeypairs == nil {
-			flightCtx.state.LocalKeypairs = make(map[elliptic.Curve]*elliptic.Keypair, len(newKeypairs))
-		}
-		maps.Copy(flightCtx.state.LocalKeypairs, newKeypairs)
-	}
-	extensions = append(extensions, &extension.KeyShare{
-		ClientShares: flightCtx.state.LocalKeyEntries,
-	})
-
-	if !slices.Contains(flightCtx.state.RemoteVersions, protocol.Version1_3) {
-		return nil, nil, dtlserrors.ErrNoCommonProtocolVersion
-	}
-	extensions = append(extensions, &extension.SupportedVersions{
-		Versions: dtlsconfig.SupportedVersionsRange(flightCtx.cfg.MinVersion, flightCtx.cfg.MaxVersion),
-	})
-
-	if len(flightCtx.cfg.LocalCertSignatureSchemes) > 0 {
-		extensions = append(extensions, &extension.SignatureAlgorithmsCert{
-			SignatureHashAlgorithms: flightCtx.cfg.LocalCertSignatureSchemes,
-		})
-	}
-
-	if len(flightCtx.cfg.ServerName) > 0 {
-		extensions = append(extensions, &extension.ServerName{ServerName: flightCtx.cfg.ServerName})
-	}
-
-	if len(flightCtx.cfg.LocalSRTPProtectionProfiles) > 0 {
-		extensions = append(extensions, &extension.UseSRTP{
-			ProtectionProfiles:  flightCtx.cfg.LocalSRTPProtectionProfiles,
-			MasterKeyIdentifier: flightCtx.cfg.LocalSRTPMasterKeyIdentifier,
-		})
-	}
-
-	if len(flightCtx.state.Cookie) > 0 {
-		extensions = append(extensions, &extension.CookieExt{Cookie: flightCtx.state.Cookie})
-	}
-
-	clientHello := &handshake.MessageClientHello{
-		Version:            protocol.Version1_2,
-		SessionID:          flightCtx.state.SessionID,
-		Cookie:             []byte{},
-		Random:             flightCtx.state.LocalRandom,
-		CipherSuiteIDs:     dtlsflight.CipherSuiteIDs(flightCtx.cfg.LocalCipherSuites),
-		CompressionMethods: dtlsflight.DefaultCompressionMethods(),
-		Extensions:         extensions,
-	}
-
-	var content handshake.Handshake
-
-	if flightCtx.cfg.ClientHelloMessageHook != nil {
-		content = handshake.Handshake{Message: flightCtx.cfg.ClientHelloMessageHook(*clientHello)}
-	} else {
-		content = handshake.Handshake{Message: clientHello}
-	}
-
-	return []*dtlsflight.Packet{
-		{
-			Record: &recordlayer.RecordLayer{
-				Header: recordlayer.Header{
-					Version: protocol.Version1_2,
-				},
-				Content: &content,
-			},
-		},
-	}, nil, nil
-}
-
-func flight5Generate(
-	_ dtlsflight.Conn,
-	flightCtx *handshakeContext,
-) ([]*dtlsflight.Packet, *alert.Alert, error) {
-	pkts, dtlsAlert, err := flight5ClientAuthPackets(flightCtx)
-	if err != nil {
-		return nil, dtlsAlert, err
-	}
-	pkts = append(pkts, HandshakePacket(&handshake.MessageFinished{}))
-	pkts[0].ResetLocalSequenceNumber = true
-
-	return pkts, nil, nil
-}
-
-func flight5ClientAuthPackets(
-	flightCtx *handshakeContext,
-) ([]*dtlsflight.Packet, *alert.Alert, error) {
-	certificateRequest, ok, err := flight5CertificateRequest(flightCtx.cache)
-	if err != nil {
-		return nil, &alert.Alert{Level: alert.Fatal, Description: alert.InternalError}, err
-	}
-	if !ok {
-		return []*dtlsflight.Packet{}, nil, nil
-	}
-
-	certificate, err := flight5ClientCertificate(flightCtx.cfg, certificateRequest)
-	if err != nil {
-		return nil, &alert.Alert{Level: alert.Fatal, Description: alert.HandshakeFailure}, err
-	}
-	if len(certificate.Certificate) == 0 {
-		return []*dtlsflight.Packet{
-			HandshakePacket(&handshake.MessageCertificate13{
-				CertificateRequestContext: append(
-					[]byte(nil),
-					certificateRequest.CertificateRequestContext...,
-				),
-			}),
-		}, nil, nil
-	}
-
-	signer, ok := certificate.PrivateKey.(crypto.Signer)
-	if !ok {
-		return nil, &alert.Alert{Level: alert.Fatal, Description: alert.HandshakeFailure},
-			dtlserrors.ErrInvalidPrivateKey
-	}
-
-	signatureScheme, err := signaturehash.SelectSignatureScheme13(
-		certificateRequestSignatureSchemes(certificateRequest),
-		signer,
-	)
-	if err != nil {
-		return nil, &alert.Alert{Level: alert.Fatal, Description: alert.InsufficientSecurity}, err
-	}
-
-	return []*dtlsflight.Packet{
-		HandshakePacket(&handshake.MessageCertificate13{
-			CertificateRequestContext: append(
-				[]byte(nil),
-				certificateRequest.CertificateRequestContext...,
-			),
-			CertificateList: certificateEntries13(certificate.Certificate),
-		}),
-		CertificateVerifyPacket(
-			&handshake.MessageCertificateVerify{
-				HashAlgorithm:      signatureScheme.Hash,
-				SignatureAlgorithm: signatureScheme.Signature,
-			},
-			signer,
-		),
-	}, nil, nil
-}
-
-func flight5CertificateRequest(
-	cache *dtlsflight.Cache,
-) (*handshake.MessageCertificateRequest13, bool, error) {
-	if cache == nil {
-		return nil, false, nil
-	}
-
-	items := cache.Pull(dtlsflight.HandshakeCachePullRule{
-		Typ:      handshake.TypeCertificateRequest,
-		Epoch:    EpochHandshake,
-		IsClient: false,
-		Optional: true,
-	})
-	if len(items) == 0 || items[0] == nil {
-		return nil, false, nil
-	}
-
-	item := items[0]
-	header := &handshake.Header{}
-	if err := header.Unmarshal(item.Data); err != nil {
-		return nil, false, err
-	}
-	if !validFlight5CertificateRequestItem(item, header) {
-		return nil, false, dtlserrors.ErrInvalidHandshakeTranscriptMessage
-	}
-
-	request := &handshake.MessageCertificateRequest13{}
-	if err := request.Unmarshal(item.Data[handshake.HeaderLength:]); err != nil {
-		return nil, false, err
-	}
-
-	return request, true, nil
-}
-
-func validFlight5CertificateRequestItem(
-	item *dtlsflight.HandshakeCacheItem,
-	header *handshake.Header,
-) bool {
-	return header.Type == handshake.TypeCertificateRequest &&
-		header.MessageSequence == item.MessageSequence &&
-		header.FragmentOffset == 0 &&
-		header.FragmentLength == header.Length &&
-		len(item.Data) == handshake.HeaderLength+int(header.Length)
-}
-
-func flight5ClientCertificate(
-	cfg *dtlsconfig.HandshakeConfig,
-	request *handshake.MessageCertificateRequest13,
-) (*tls.Certificate, error) {
-	requestInfo := &dtlsconfig.CertificateRequestInfo{
-		SignatureSchemes: certificateRequestSignatureSchemes(request),
-	}
-	for _, ext := range request.Extensions {
-		if authorities, ok := ext.(*extension.CertificateAuthorities); ok {
-			requestInfo.AcceptableCAs = authorities.Authorities
-
-			break
-		}
-	}
-
-	certificate, err := cfg.GetClientCertificate(requestInfo)
-	if err != nil {
-		return nil, err
-	}
-	if certificate == nil {
-		return &tls.Certificate{}, nil
-	}
-
-	return certificate, nil
-}
-
-func certificateRequestSignatureSchemes(
-	request *handshake.MessageCertificateRequest13,
-) []signaturehash.Algorithm {
-	for _, ext := range request.Extensions {
-		if algorithms, ok := ext.(*extension.SupportedSignatureAlgorithms); ok {
-			return algorithms.SignatureHashAlgorithms
-		}
-	}
-
-	return nil
-}
-
-func certificateEntries13(certificates [][]byte) []handshake.CertificateEntry13 {
-	entries := make([]handshake.CertificateEntry13, 0, len(certificates))
-	for _, certificate := range certificates {
-		entries = append(entries, handshake.CertificateEntry13{
-			CertificateData: append([]byte(nil), certificate...),
-		})
-	}
-
-	return entries
-}
